requirements.txt
pyyaml
requests
coloredlogs
